requirements.txt
Django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.4
pillow>=11.0
python-docx>=1.1
openpyxl>=3.1
